requirements.txt
# Core Flask
Flask==2.3.3

# ORM & DB
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0

# Migrations
Flask-Migrate==4.0.5

# Authentification JWT
Flask-JWT-Extended==4.6.0

# Sérialisation / Validation
marshmallow==3.21.1
Flask-Marshmallow==0.15.0

# Chargement des variables d’environnement
python-dotenv==1.0.1

# Utilitaires CLI
Flask-Script==2.0.6

# Export PDF/Excel/CSV
reportlab==4.0.8
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1

# Sécurité (hash de mot de passe)
Werkzeug==2.3.7

cryptography

Flask-Migrate

marshmallow-sqlalchemy
